C#: Some re-factoring of NugetPackages and logic for file downloading.
